Exposed the server creation form based on Admin configuration.
This is just UI and a bit of backend smarts to make the UI responsive / reactive.  This is not complete, and the user (at this time) would desimate the current main server if he / she were to create one.  Don't use this version of the Vector software if you are putting it into any level of production for consistent use.